requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
beautifulsoup4==4.4.1
billiard==3.3.0.23
bs4==0.0.1
celery==3.1.25
click==6.6
decorator==4.0.6
Django==1.9.1
django-activity-stream==0.6.3
django-appconf==1.0.1
django-axes==2.0.0
django-celery==3.2.1
django-configurations==2.0
django-digg-paginator==0.1.3
django-mptt==0.8.7
django-picklefield==0.3.2
django-redis-cache==1.6.5
django-rq==0.9.1
django-user-accounts==2.0.0
django-widget-tweaks==1.4.1
djangorestframework==3.5.4
EasyProcess==0.2.3
ecdsa==0.13
enum34==1.1.6
Fabric==1.11.1
fabric-virtualenv==0.3.0
fabvenv==0.2.1
fpdf==1.7.2
fuzzywuzzy==0.12.0
geoip2==2.3.0
gnureadline==6.3.3
googlemaps==2.4.4
httplib2==0.10.3
ipaddr==2.1.11
ipaddress==1.0.18
ipython==4.0.1
ipython-genutils==0.1.0
jsonfield==1.0.3
kombu==3.0.37
maxminddb==1.2.0
mechanize==0.2.5
MySQL-python==1.2.5
numpy==1.10.4
oauth2client==4.0.0
olefile==0.44
pandas==0.19.2
paramiko==1.17.0
path.py==8.1.2
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.13
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycrypto==2.6.1
Pygments==2.2.0
pymongo==3.2
pyPdf==1.13
PyPDF2==1.26.0
python-dateutil==2.4.2
python-Levenshtein==0.12.0
pytz==2015.7
PyVirtualDisplay==0.2.1
redis==2.10.5
requests==2.12.4
rq==0.6.0
rsa==3.4.2
scandir==1.4
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.5
tablib==0.11.1
traitlets==4.0.0
uritemplate==3.0.0
wcwidth==0.1.7
xlrd==1.0.0
XlsxWriter==0.9.6
xmltodict==0.10.1
